refactor: improve villager profession test suite
